refactor(core): extract pipeline bind and scale-training modules
Peel data/layer binding and axis/color scale training out of the
pipeline orchestrator into pipeline/bind.ts and pipeline/scale-training.ts.
Public runPipeline surface is unchanged. Characterization tests cover
bind error codes, axis type inference, and color scale resolution.
