+++ packages/core/src/pipeline.ts
@@ -36,20 +36,16 @@
  * values with a warning and REFUSE non-positive explicit domains.
  */
 import type {
-  Aes,
   BarParams,
   BoxplotParams,
-  ChannelValue,
-  ColorScaleSpec,
   ErrorbarParams,
-  LayerSpec,
   PortableSpec,
   PositionParams,
   PositionScaleSpec,
   SmoothParams,
   SpecInput,
 } from "@ggsvelte/spec";
-import { didYouMean, normalize, SpecValidationError, validate } from "@ggsvelte/spec";
+import { normalize, SpecValidationError, validate } from "@ggsvelte/spec";
 
 import { deriveGroups } from "./grouping.js";
 import { FONT_METRICS } from "./layout/font-metrics.js";
@@ -62,19 +58,9 @@ import { defaultLogTickFormat } from "./layout/ticks.js";
 import { defaultTimeTickFormat } from "./layout/time.js";
 import type { LegendInput, LegendOrder } from "./legend.js";
 import { buildLegends } from "./legend.js";
-import { trainSequential, VIRIDIS_RAMP_10 } from "./scales/color.js";
 import type { ScaleState } from "./scales/state.js";
-import { PaletteExhaustedError } from "./scales/state.js";
-import type { ColorScale, ContinuousConfig, PositionScale } from "./scales/train.js";
-import {
-  bandKey,
-  CATEGORICAL_PALETTE_10,
-  finiteExtent,
-  ScaleConfigError,
-  trainBand,
-  trainColor,
-  trainContinuous,
-} from "./scales/train.js";
+import type { PositionScale } from "./scales/train.js";
+import { bandKey, finiteExtent } from "./scales/train.js";
 import type {
   GeometryBatch,
   Scene,
@@ -92,9 +78,8 @@ import { statSmooth } from "./stats/smooth.js";
 import { statSummary } from "./stats/summary.js";
 import { DEFAULT_JITTER_SEED, jitterOffsets, nudgeOffsets } from "./positions/jitter.js";
 import { positionDodge, positionStack } from "./positions/positions.js";
-import type { CellValue, Columns, Discreteness, Rows } from "./table.js";
+import type { CellValue, Discreteness } from "./table.js";
 import { cellsToNumeric, cellToNumber, ColumnTable } from "./table.js";
-import type { EditionDefaults } from "./editions.js";
 import { resolveEditionDefaults } from "./editions.js";
 import type { ThemeTokens } from "./theme.js";
 import { resolveTheme, UnknownThemeError } from "./theme.js";
@@ -110,16 +95,13 @@ import { LineageStore } from "./identity.js";
 import type {
   Advisory,
   AxisValueFormatter,
-  ColorBinding,
   LayerBackend,
   LayerBinding,
   LayerFrame,
   MappedField,
-  NamedData,
   PipelineWarning,
   RenderModel,
   ResolvedColorScale,
-  RuleForm,
   RunOptions,
   ScaleDomainSnapshot,
 } from "./pipeline/types.js";
@@ -127,6 +109,13 @@ import { CANVAS_AUTO_THRESHOLD, NO_ROW, PipelineError } from "./pipeline/types.j
 import { batchMarkCount, buildBatch, flipBatchInPlace } from "./pipeline/geometry.js";
 import type { Frame } from "./pipeline/geometry.js";
 import { resolveFacet, SINGLE_PANEL } from "./pipeline/facets.js";
+import { bindData, bindLayer } from "./pipeline/bind.js";
+import {
+  collectAxisInputs,
+  isBarLike,
+  resolveColorScale,
+  trainAxis,
+} from "./pipeline/scale-training.js";
 
 // Re-export the public pipeline contract (import path stability).
 export type {
@@ -144,362 +133,6 @@ export type {
 } from "./pipeline/types.js";
 export { CANVAS_AUTO_THRESHOLD, PipelineError } from "./pipeline/types.js";
 export { batchMarkCount } from "./pipeline/geometry.js";
-
-// ---------------------------------------------------------------------------
-// Data binding
-// ---------------------------------------------------------------------------
-
-function tableFromNamed(data: NamedData): ColumnTable {
-  if (Array.isArray(data)) return ColumnTable.fromRows(data);
-  if ("values" in data && Array.isArray((data as { values: unknown }).values)) {
-    return ColumnTable.fromRows((data as { values: Rows }).values);
-  }
-  if ("columns" in data && typeof (data as { columns: unknown }).columns === "object") {
-    return ColumnTable.fromColumns((data as { columns: Columns }).columns);
-  }
-  return ColumnTable.fromColumns(data as Columns);
-}
-
-function bindData(spec: PortableSpec, options: RunOptions): ColumnTable {
-  const ref = spec.data;
-  if (ref === undefined) {
-    throw new PipelineError(
-      "no-data",
-      "/data",
-      "The spec has no data. Provide spec.data ({values}, {columns}, or {name}) or layer data.",
-    );
-  }
-  if ("values" in ref) return ColumnTable.fromRows(ref.values);
-  if ("columns" in ref) return ColumnTable.fromColumns(ref.columns);
-  const name = ref.name;
-  const fromSpec = spec.datasets?.[name];
-  const fromRun = options.data?.[name];
-  if (fromSpec !== undefined && fromRun !== undefined && options.allowOverride !== true) {
-    throw new PipelineError(
-      "dataset-collision",
-      "/data/name",
-      `Dataset "${name}" is defined in both spec.datasets and RunOptions.data. ` +
-        "Pass allowOverride: true to let the runtime data win, or rename one.",
-    );
-  }
-  if (fromRun !== undefined && (fromSpec === undefined || options.allowOverride === true)) {
-    return tableFromNamed(fromRun);
-  }
-  if (fromSpec !== undefined) return tableFromNamed(fromSpec);
-  const available = [...Object.keys(spec.datasets ?? {}), ...Object.keys(options.data ?? {})];
-  throw new PipelineError(
-    "unknown-dataset",
-    "/data/name",
-    `Unknown dataset "${name}". Available: ${available.length > 0 ? available.join(", ") : "none"}.`,
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Layer binding (channel resolution + structural rules)
-// ---------------------------------------------------------------------------
-
-/** y-channel { stat } columns each stat exposes (module-header contracts). */
-const STAT_Y_COLUMNS: Record<string, readonly string[]> = {
-  identity: [],
-  count: ["count"],
-  bin: ["count", "density", "ncount", "ndensity"],
-  density: ["density", "count", "scaled", "ndensity"],
-  smooth: [],
-  boxplot: [],
-  summary: [],
-};
-
-function checkField(
-  channel: ChannelValue | undefined,
-  channelName: string,
-  layerIndex: number,
-  table: ColumnTable,
-  warnings: PipelineWarning[],
-): string | null {
-  if (channel === undefined || channel === null) return null;
-  if (!("field" in channel)) return null;
-  if (!table.has(channel.field)) {
-    const suggestion = didYouMean(channel.field, table.fields);
-    throw new PipelineError(
-      "unknown-field",
-      `/layers/${layerIndex}/aes/${channelName}`,
-      `Unknown field "${channel.field}" (available: ${table.fields.join(", ") || "none"}).` +
-        (suggestion === undefined ? "" : ` Did you mean "${suggestion}"?`),
-    );
-  }
-  if (allNull(table.column(channel.field))) {
-    throw new PipelineError(
-      "all-null-column",
-      `/layers/${layerIndex}/aes/${channelName}`,
-      `Field "${channel.field}" contains only null values; the "${channelName}" channel cannot be drawn from it.`,
-    );
-  }
-  void warnings;
-  return channel.field;
-}
-
-function allNull(column: readonly CellValue[]): boolean {
-  if (column.length === 0) return false;
-  for (const v of column) if (v !== null) return false;
-  return true;
-}
-
-function requireField(
-  field: string | null,
-  channelName: string,
-  layerIndex: number,
-  geom: string,
-): string {
-  if (field === null) {
-    throw new PipelineError(
-      "missing-channel",
-      `/layers/${layerIndex}/aes/${channelName}`,
-      `The ${geom} geom requires a "${channelName}" channel (map it with aes).`,
-    );
-  }
-  return field;
-}
-
-function colorBinding(
-  channel: ChannelValue | undefined,
-  channelName: string,
-  layerIndex: number,
-  table: ColumnTable,
-  warnings: PipelineWarning[],
-): ColorBinding {
-  const out: ColorBinding = { field: null, constant: null, scaledConstant: null };
-  if (channel === undefined || channel === null) return out;
-  if ("field" in channel) {
-    out.field = checkField(channel, channelName, layerIndex, table, warnings);
-  } else if ("value" in channel) {
-    if (channel.scale === true) out.scaledConstant = channel.value;
-    else out.constant = String(channel.value);
-  } else {
-    warnings.push({
-      code: "stat-channel-unsupported",
-      message: `Layer ${layerIndex}: { stat } mappings on the "${channelName}" channel are not supported yet; the mapping is ignored.`,
-    });
-  }
-  return out;
-}
-
-function bindLayer(
-  layer: LayerSpec,
-  index: number,
-  table: ColumnTable,
-  warnings: PipelineWarning[],
-): LayerBinding {
-  const aes: Aes = layer.aes ?? {};
-  const geom = layer.geom;
-  const params = layer.params ?? {};
-
-  // --- rule: two honest forms ------------------------------------------------
-  let ruleForm: RuleForm | null = null;
-  if (geom === "rule") {
-    const p = params as { xintercept?: unknown; yintercept?: unknown };
-    const hasIntercepts = p.xintercept !== undefined || p.yintercept !== undefined;
-    const xMapped = aes.x !== undefined && aes.x !== null;
-    const yMapped = aes.y !== undefined && aes.y !== null;
-    if (hasIntercepts && (xMapped || yMapped)) {
-      throw new PipelineError(
-        "rule-form-ambiguous",
-        `/layers/${index}`,
-        "This rule layer mixes the annotation form (params.xintercept/yintercept) with mapped aes.x/aes.y. Use fixed intercepts OR a data mapping, never both.",
-      );
-    }
-    if (!hasIntercepts && !xMapped && !yMapped) {
-      throw new PipelineError(
-        "rule-form-missing",
-        `/layers/${index}`,
-        "This rule layer has neither fixed intercepts (params.xintercept/yintercept) nor a mapped aes.x/aes.y — nothing to draw.",
-      );
-    }
-    if (!hasIntercepts && xMapped && yMapped) {
-      throw new PipelineError(
-        "rule-both-axes",
-        `/layers/${index}`,
-        "This rule layer maps BOTH aes.x and aes.y; a data-driven rule is either vertical (map x) or horizontal (map y). Unset the other channel with null.",
-      );
-    }
-    ruleForm = hasIntercepts ? "annotation" : xMapped ? "vertical" : "horizontal";
-  }
-
-  const stat = layer.stat ?? "identity";
-  const xField = checkField(aes.x, "x", index, table, warnings);
-  let yField: string | null = null;
-  let yStatColumn: string | null = null;
-  const y = aes.y;
-  if (y !== undefined && y !== null && "stat" in y) {
-    const generated = STAT_Y_COLUMNS[stat] ?? [];
-    if (!generated.includes(y.stat)) {
-      throw new PipelineError(
-        "unknown-stat-column",
-        `/layers/${index}/aes/y`,
-        `Channel "y" maps stat column "${y.stat}", but this layer's stat ("${stat}") ${generated.length > 0 ? `generates: ${generated.join(", ")}` : "generates no y-mappable columns"}.`,
-      );
-    }
-    yStatColumn = y.stat;
-  } else {
-    yField = checkField(y, "y", index, table, warnings);
-  }
-
-  if (geom === "bar" && yField !== null) {
-    throw new PipelineError(
-      "computed-y-mapped",
-      `/layers/${index}/aes/y`,
-      `The bar geom computes y with the ${stat} stat, so aes.y must not map data. Use geom "col" for pre-computed heights.`,
-    );
-  }
-  if (geom === "density" && yField !== null) {
-    throw new PipelineError(
-      "computed-y-mapped",
-      `/layers/${index}/aes/y`,
-      "The density geom computes y with the density stat, so aes.y must not map data. Map only x.",
-    );
-  }
-  if (stat === "bin") {
-    const p = params as BarParams;
-    if (p.center !== undefined && p.boundary !== undefined) {
-      throw new PipelineError(
-        "bin-center-and-boundary",
-        `/layers/${index}/params`,
-        "The bin stat accepts params.center OR params.boundary (both align the bin grid), never both.",
-      );
-    }
-    if (xField !== null && table.fieldType(xField) === "nominal") {
-      throw new PipelineError(
-        "channel-type-mismatch",
-        `/layers/${index}/aes/x`,
-        `The bin stat needs a continuous x, but field "${xField}" is nominal. Use geom "bar" (the count stat) to count categories instead.`,
-      );
-    }
-  }
-  if (geom === "density" && xField !== null && table.fieldType(xField) === "nominal") {
-    throw new PipelineError(
-      "channel-type-mismatch",
-      `/layers/${index}/aes/x`,
-      `The density stat needs a continuous x, but field "${xField}" is nominal. Use geom "bar" (the count stat) to count categories instead.`,
-    );
-  }
-  if (geom === "smooth") {
-    for (const [channel, field] of [
-      ["x", xField],
-      ["y", yField],
-    ] as const) {
-      if (field !== null && table.fieldType(field) === "nominal") {
-        throw new PipelineError(
-          "channel-type-mismatch",
-          `/layers/${index}/aes/${channel}`,
-          `The smooth stat needs quantitative x and y, but field "${field}" (${channel}) is nominal.`,
-        );
-      }
-    }
-  }
-  if (geom === "boxplot") {
-    if (xField !== null && table.fieldType(xField) !== "nominal") {
-      throw new PipelineError(
-        "channel-type-mismatch",
-        `/layers/${index}/aes/x`,
-        `The boxplot geom needs a DISCRETE x this milestone, but field "${xField}" is ${table.fieldType(xField)}. Map x to a categorical field.`,
-      );
-    }
-    if (yField !== null && table.fieldType(yField) === "nominal") {
-      throw new PipelineError(
-        "channel-type-mismatch",
-        `/layers/${index}/aes/y`,
-        `The boxplot stat needs a quantitative y, but field "${yField}" is nominal.`,
-      );
-    }
-  }
-
-  // --- ymin/ymax (errorbar identity form) --------------------------------------
-  const yminField = checkField(aes.ymin, "ymin", index, table, warnings);
-  const ymaxField = checkField(aes.ymax, "ymax", index, table, warnings);
-
-  // --- required channels ------------------------------------------------------
-  if (
-    geom === "point" ||
-    geom === "line" ||
-    geom === "col" ||
-    geom === "area" ||
-    geom === "text" ||
-    geom === "smooth" ||
-    geom === "boxplot"
-  ) {
-    requireField(xField, "x", index, geom);
-    if (yStatColumn === null) requireField(yField, "y", index, geom);
-  }
-  if (geom === "bar" || geom === "density") requireField(xField, "x", index, geom);
-  if (geom === "errorbar") {
-    requireField(xField, "x", index, geom);
-    if (stat === "summary") {
-      requireField(yField, "y", index, geom);
-    } else {
-      requireField(yminField, "ymin", index, geom);
-      requireField(ymaxField, "ymax", index, geom);
-    }
-  }
-  if (geom === "rule" && ruleForm === "vertical") requireField(xField, "x", index, geom);
-  if (geom === "rule" && ruleForm === "horizontal") requireField(yField, "y", index, geom);
-
-  // --- label / weight ----------------------------------------------------------
-  let labelField: string | null = null;
-  let labelConstant: string | null = null;
-  const label = aes.label;
-  if (label !== undefined && label !== null) {
-    if ("field" in label) labelField = checkField(label, "label", index, table, warnings);
-    else if ("value" in label) labelConstant = String(label.value);
-  }
-  if (geom === "text" && labelField === null && labelConstant === null) {
-    throw new PipelineError(
-      "missing-channel",
-      `/layers/${index}/aes/label`,
-      'The text geom requires a "label" channel (map it with aes).',
-    );
-  }
-  const weightField = checkField(aes.weight, "weight", index, table, warnings);
-
-  const color = colorBinding(aes.color, "color", index, table, warnings);
-  const fill = colorBinding(aes.fill, "fill", index, table, warnings);
-  if (
-    (geom === "bar" ||
-      geom === "col" ||
-      geom === "area" ||
-      geom === "boxplot" ||
-      geom === "density") &&
-    (color.field !== null || color.constant !== null || color.scaledConstant !== null)
-  ) {
-    warnings.push({
-      code: "color-on-fill-geom",
-      message: `Layer ${index} (${geom}): the color channel styles OUTLINES, which this geom does not support as a data channel yet — map "fill" instead. The color mapping is ignored.`,
-    });
-    color.field = null;
-    color.constant = null;
-    color.scaledConstant = null;
-  }
-  if (weightField !== null && (stat === "boxplot" || stat === "smooth" || stat === "summary")) {
-    warnings.push({
-      code: "weight-unsupported",
-      message: `Layer ${index}: the ${stat} stat does not support aes.weight; the weight mapping is ignored.`,
-    });
-  }
-
-  return {
-    layer,
-    index,
-    xField,
-    yField,
-    yStatColumn,
-    yminField,
-    ymaxField,
-    color,
-    fill,
-    labelField,
-    labelConstant,
-    weightField,
-    ruleForm,
-  };
-}
 
 // ---------------------------------------------------------------------------
 // Layer frames (post-stat, post-position data)
@@ -948,11 +581,6 @@ function buildFrame(
   };
 }
 
-/** Geoms whose marks extend from a zero baseline on the measure (y) axis. */
-function isBarLike(geom: LayerSpec["geom"]): boolean {
-  return geom === "bar" || geom === "col" || geom === "area";
-}
-
 /** Per-row slot keys: band categories, or bin centers for binned bars. */
 function slotKeys(frame: LayerFrame): (number | string)[] | null {
   if (frame.xValues !== null) return frame.xValues.map((v) => bandKey(v));
@@ -1041,400 +669,6 @@ function applyPosition(frame: LayerFrame, advisories: Advisory[], table: ColumnT
     frame.dodgeSlot = dodge.slot;
     frame.dodgeSlotCounts = dodge.slotCount;
   }
-}
-
-// ---------------------------------------------------------------------------
-// Scale configuration + training
-// ---------------------------------------------------------------------------
-
-const POSITION_TYPE_OVERRIDE =
-  'Set scales.AXIS.type ("linear" | "log" | "time" | "band") in the spec.';
-
-function continuousDomainOf(
-  config: PositionScaleSpec | undefined,
-  axis: "x" | "y",
-): [number, number] | undefined {
-  if (config?.domain === undefined) return undefined;
-  if (config.domain.length !== 2) {
-    throw new PipelineError(
-      "invalid-scale-domain",
-      `/scales/${axis}/domain`,
-      `A continuous ${axis} domain needs exactly [min, max] (got ${config.domain.length} entries).`,
-    );
-  }
-  const lo = cellToNumber(config.domain[0] as CellValue);
-  const hi = cellToNumber(config.domain[1] as CellValue);
-  if (!Number.isFinite(lo) || !Number.isFinite(hi)) {
-    throw new PipelineError(
-      "invalid-scale-domain",
-      `/scales/${axis}/domain`,
-      `The ${axis} domain [${String(config.domain[0])}, ${String(config.domain[1])}] does not parse to finite numbers (use numbers, or ISO 8601 strings for time scales).`,
-    );
-  }
-  return lo <= hi ? [lo, hi] : [hi, lo];
-}
-
-interface AxisTraining {
-  scale: PositionScale;
-  advisories: Advisory[];
-  warnings: PipelineWarning[];
-}
-
-interface AxisInputs {
-  /** Discrete columns contributing to a band domain. */
-  columns: (readonly CellValue[])[];
-  /** Continuous numeric arrays (post-stat / post-position). */
-  numeric: Float64Array[];
-  /** True when any contributing field is discrete or a geom forces bands. */
-  anyDiscrete: boolean;
-  /** True when every contributing field is temporal (and none discrete). */
-  allTemporal: boolean;
-  /** True when a bar-like geom measures on this axis (zero forcing). */
-  barMeasure: boolean;
-  /** Human-readable description of the evidence (advisory text). */
-  evidence: string;
-}
-
-function trainAxis(
-  axis: "x" | "y",
-  inputs: AxisInputs,
-  config: PositionScaleSpec | undefined,
-): AxisTraining {
-  const advisories: Advisory[] = [];
-  const warnings: PipelineWarning[] = [];
-  const howToOverride = POSITION_TYPE_OVERRIDE.replace("AXIS", axis);
-
-  let type = config?.type;
-  if (type === undefined) {
-    type = inputs.anyDiscrete ? "band" : inputs.allTemporal ? "time" : "linear";
-    advisories.push({
-      code: "scale-type-inferred",
-      path: `scales.${axis}`,
-      chosen: `${type} (${inputs.evidence})`,
-      howToOverride,
-    });
-  }
-
-  if (type === "band") {
-    const domain = config?.domain;
-    const scale = trainBand(inputs.columns, {
-      ...(domain !== undefined && { domain }),
-      ...(config?.reverse !== undefined && { reverse: config.reverse }),
-    });
-    if (scale.domain.length === 0) {
-      warnings.push({
-        code: "empty-domain",
-        message: `The ${axis} band scale has no categories.`,
-      });
-    }
-    return { scale, advisories, warnings };
-  }
-
-  // --- continuous ------------------------------------------------------------
-  let zero = config?.zero;
-  if (inputs.barMeasure && type !== "time" && zero === undefined && config?.domain === undefined) {
-    zero = true;
-    advisories.push({
-      code: "zero-forced",
-      path: `scales.${axis}`,
-      chosen: "domain extended to include 0 (bars/areas measure from a zero baseline)",
-      howToOverride: `Set scales.${axis}.zero to false, or pin scales.${axis}.domain.`,
-    });
-  }
-
-  const domain = continuousDomainOf(config, axis);
-  const continuousConfig: ContinuousConfig = {
-    type,
-    ...(domain !== undefined && { domain }),
-    ...(config?.nice !== undefined && { nice: config.nice }),
-    ...(zero !== undefined && { zero }),
-    ...(config?.reverse !== undefined && { reverse: config.reverse }),
-  };
-  let training;
-  try {
-    training = trainContinuous(inputs.numeric, continuousConfig);
-  } catch (error) {
-    if (error instanceof ScaleConfigError) {
-      throw new PipelineError(error.code, `/scales/${axis}`, error.message);
-    }
-    throw error;
-  }
-  if (training.empty) {
-    warnings.push({
-      code: "empty-domain",
-      message: `The ${axis} scale has no finite${type === "log" ? " positive" : ""} values; using a default domain.`,
-    });
-  }
-  if (training.nonPositive > 0) {
-    warnings.push({
-      code: "log-nonpositive",
-      message: `Removed ${training.nonPositive} non-positive value(s) from the ${axis} log scale (log10 is undefined at or below zero).`,
-    });
-  }
-  return { scale: training.scale, advisories, warnings };
-}
-
-/** Collect per-axis training evidence across layers. */
-function collectAxisInputs(
-  axis: "x" | "y",
-  frames: readonly LayerFrame[],
-  configType: PositionScaleSpec["type"] | undefined,
-  advisories: Advisory[],
-): AxisInputs {
-  const columns: (readonly CellValue[])[] = [];
-  const numeric: Float64Array[] = [];
-  let anyDiscrete = false;
-  let allTemporal = true;
-  let sawContinuousEvidence = false;
-  let barMeasure = false;
-  const typeParts = new Set<string>();
-
-  for (const frame of frames) {
-    const { binding } = frame;
-    const geom = binding.layer.geom;
-
-    if (axis === "x") {
-      if (frame.xmin !== null && frame.xmax !== null) {
-        // Binned bars: the x domain is the union of bin edges (continuous).
-        numeric.push(frame.xmin, frame.xmax);
-        const field = binding.xField!;
-        const fieldType = frame.table.has(field) ? frame.table.fieldType(field) : "quantitative";
-        typeParts.add(`binned ${fieldType}`);
-        if (fieldType !== "temporal") allTemporal = false;
-        sawContinuousEvidence = true;
-      } else if (frame.xValues !== null || frame.xNumeric !== null) {
-        if (frame.xValues !== null) columns.push(frame.xValues);
-        if (frame.xNumeric !== null) numeric.push(frame.xNumeric);
-        const field = binding.xField!;
-        const fieldType = frame.table.has(field) ? frame.table.fieldType(field) : "quantitative";
-        typeParts.add(fieldType);
-        const barX = (geom === "bar" || geom === "col") && binding.layer.stat !== "bin";
-        if (barX && fieldType !== "nominal" && configType === undefined) {
-          advisories.push({
-            code: "bar-x-discretized",
-            path: `layers.${binding.index}`,
-            chosen: `x treated as discrete bands (${geom} geom)`,
-            howToOverride: "Use point/line/area for continuous x, or set scales.x.type explicitly.",
-          });
-        }
-        if (barX || fieldType === "nominal") anyDiscrete = true;
-        if (fieldType !== "temporal") allTemporal = false;
-        sawContinuousEvidence = true;
-      }
-      if (frame.box !== null && frame.box.outlierX.length > 0) {
-        columns.push(frame.box.outlierX);
-      }
-      for (const v of frame.xIntercepts) {
-        columns.push([v]);
-        numeric.push(Float64Array.of(cellToNumber(v)));
-        if (typeof v === "string" && !Number.isFinite(cellToNumber(v))) anyDiscrete = true;
-        sawContinuousEvidence = true;
-      }
-    } else {
-      if (isBarLike(geom) || geom === "density") barMeasure = true;
-      if (frame.ymin !== null && frame.ymax !== null) {
-        numeric.push(frame.ymin, frame.ymax);
-        // Bands need not cover the center line (se: false smooths have
-        // NaN bands; the summary center can escape min/max bounds).
-        if ((geom === "smooth" || geom === "errorbar") && frame.yNumeric !== null) {
-          numeric.push(frame.yNumeric);
-        }
-        if (frame.box !== null) numeric.push(frame.box.outlierY);
-        typeParts.add("quantitative");
-        allTemporal = false;
-        sawContinuousEvidence = true;
-      } else if (binding.yStatColumn !== null && frame.yNumeric !== null) {
-        numeric.push(frame.yNumeric);
-        typeParts.add(binding.yStatColumn);
-        allTemporal = false;
-        sawContinuousEvidence = true;
-      } else if (binding.yField !== null) {
-        // Panel-local data: free-y facets train each panel on ITS rows.
-        const column = frame.table.column(binding.yField);
-        columns.push(column);
-        numeric.push(frame.table.numeric(binding.yField));
-        const fieldType = frame.table.fieldType(binding.yField);
-        typeParts.add(fieldType);
-        if (fieldType === "nominal") anyDiscrete = true;
-        if (fieldType !== "temporal") allTemporal = false;
-        sawContinuousEvidence = true;
-      }
-      for (const v of frame.yIntercepts) {
-        columns.push([v]);
-        numeric.push(Float64Array.of(cellToNumber(v)));
-        sawContinuousEvidence = true;
-      }
-    }
-  }
-
-  return {
-    columns,
-    numeric,
-    anyDiscrete,
-    allTemporal: allTemporal && sawContinuousEvidence,
-    barMeasure,
-    evidence: `field type: ${[...typeParts].join("+") || "none"}`,
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Color / fill scale resolution
-// ---------------------------------------------------------------------------
-
-interface ColorResolution {
-  resolved: ResolvedColorScale | null;
-  legendInput: LegendInput | null;
-  state: ScaleState | null;
-}
-
-function resolveColorScale(
-  name: "color" | "fill",
-  frames: readonly LayerFrame[],
-  table: ColumnTable,
-  config: ColorScaleSpec | undefined,
-  prevState: ScaleState | null,
-  legendTitle: string,
-  warnings: PipelineWarning[],
-  advisories: Advisory[],
-  editionDefaults: EditionDefaults,
-): ColorResolution {
-  const values: CellValue[] = [];
-  let anyDiscreteField = false;
-  let anyField = false;
-  for (const frame of frames) {
-    const channel = name === "color" ? frame.binding.color : frame.binding.fill;
-    const frameValues = name === "color" ? frame.colorValues : frame.fillValues;
-    if (channel.field !== null && frameValues !== null) {
-      anyField = true;
-      if (table.has(channel.field) && table.discreteness(channel.field) === "discrete") {
-        anyDiscreteField = true;
-      }
-      for (const v of frameValues) values.push(v);
-    }
-    if (channel.scaledConstant !== null) {
-      anyDiscreteField = true;
-      anyField = true;
-      values.push(channel.scaledConstant);
-    }
-  }
-  if (!anyField) return { resolved: null, legendInput: null, state: null };
-
-  const type = config?.type ?? (anyDiscreteField ? "ordinal" : "sequential");
-
-  if (type === "sequential") {
-    if (anyDiscreteField) {
-      warnings.push({
-        code: "sequential-discrete-field",
-        message: `The ${name} scale is sequential but a mapped field is discrete; values that do not parse as numbers render the unknown color.`,
-      });
-    }
-    const numeric = cellsToNumeric(values);
-    const extent = finiteExtent([numeric]);
-    const domain = config?.domain;
-    const sequentialDomain =
-      domain !== undefined && domain.length === 2
-        ? ([cellToNumber(domain[0] as CellValue), cellToNumber(domain[1] as CellValue)] as [
-            number,
-            number,
-          ])
-        : undefined;
-    // Edition-keyed default ramp: identical to the trainSequential built-in
-    // for edition 1 (pass nothing — keeps behavior byte-stable); a different
-    // edition's ramp is passed explicitly. Explicit config always wins.
-    const editionRamp =
-      editionDefaults.sequentialRamp === VIRIDIS_RAMP_10
-        ? undefined
-        : editionDefaults.sequentialRamp;
-    const range = config?.range ?? editionRamp;
-    const scale = trainSequential(extent, {
-      ...(sequentialDomain !== undefined && { domain: sequentialDomain }),
-      ...(range !== undefined && { range }),
-      ...(config?.reverse !== undefined && { reverse: config.reverse }),
-    });
-    if (config?.scheme === undefined && config?.range === undefined) {
-      advisories.push({
-        code: "palette-inferred",
-        path: `scales.${name}`,
-        chosen: "sequential viridis ramp",
-        howToOverride: `Set scales.${name}.range (ramp stops) or scales.${name}.domain.`,
-      });
-    }
-    const labelFormat = config?.labels;
-    let format = defaultTickFormat(tickStep(scale.domain[0], scale.domain[1], 5));
-    if (labelFormat !== undefined) {
-      const f = numberFormatter(labelFormat);
-      if (f.ok) {
-        format = (v: number) => f.format(v);
-      } else {
-        warnings.push({
-          code: "invalid-label-format",
-          message: `Unrecognized labels format "${labelFormat}" on scales.${name}; using the default.`,
-        });
-      }
-    }
-    return {
-      resolved: { kind: "sequential", scale },
-      legendInput: {
-        kind: "ramp",
-        scale: name,
-        title: legendTitle,
-        domain: scale.domain,
-        at: (t: number) => scale.at(t),
-        format,
-      },
-      state: null,
-    };
-  }
-
-  // --- ordinal (value-stable) --------------------------------------------------
-  const scheme = config?.scheme;
-  // Edition-keyed default palette: for edition 1 nothing is passed (trainColor
-  // keeps its "observable10" scheme fingerprint — byte-stable with pre-edition
-  // state); other editions pass their palette as an explicit range.
-  const editionPalette =
-    editionDefaults.categoricalPalette === CATEGORICAL_PALETTE_10
-      ? undefined
-      : editionDefaults.categoricalPalette;
-  // A named scheme resolves inside trainColor. Edition defaults only apply
-  // when the caller supplied neither a scheme nor an explicit range.
-  const range = config?.range ?? (scheme === undefined ? editionPalette : undefined);
-  let scale: ColorScale;
-  try {
-    scale = trainColor(values, prevState, {
-      ...(config?.domain !== undefined && { domain: config.domain }),
-      ...(config?.domainMode !== undefined && { domainMode: config.domainMode }),
-      ...(range !== undefined && { range }),
-      ...(scheme !== undefined && { scheme }),
-      ...(config?.reverse !== undefined && { reverse: config.reverse }),
-      ...(config?.onExhaust !== undefined && { onExhaust: config.onExhaust }),
-    });
-  } catch (error) {
-    if (error instanceof PaletteExhaustedError) {
-      throw new PipelineError("palette-exhausted", `/scales/${name}`, error.message);
-    }
-    throw error;
-  }
-  for (const w of scale.warnings) warnings.push({ code: w.code, message: w.message });
-  if (config?.scheme === undefined && config?.range === undefined) {
-    advisories.push({
-      code: "palette-inferred",
-      path: `scales.${name}`,
-      chosen: "categorical 10-color palette (value-stable assignment)",
-      howToOverride: `Set scales.${name}.scheme, scales.${name}.range, or scales.${name}.domain.`,
-    });
-  }
-  return {
-    resolved: { kind: "ordinal", scale },
-    legendInput: {
-      kind: "discrete",
-      scale: name,
-      title: legendTitle,
-      domain: scale.domain,
-      firstSeen: values,
-      colorOf: (v: unknown) => scale.colorOf(v),
-    },
-    state: scale.state,
-  };
 }
 
 // ---------------------------------------------------------------------------
